Cleaning up the "dist" constant handling system.
Removing the mess with methods and nesting and all that. Interestingly,
Teensy>=3 has a single address space and progmem is also not neccessary.
That simplifies a lot of things. The storage is now just a simple key-value
dict just like some unix environment dictionary.

Also introducing some versioning with this commit.