Some code for computing an in-situ checksum from the program image.
This is useful for letting the future plugin system working without
having a precise notion of versions. That is, if one can assert that a
specific firmware image is running on a teensy, at runtime, this is
better then any versioning. Requires reproducable builds, thought. I
guess the pio setup provides that.

Anyway, the code just computes a sha256 sum of the firmware image which
is correct. And it is fast.

The code compiles, has been somewhat tested (quick and dirty within the
main function) but not at the message handler integration as provided
in this commit, since my current development teensy lacks internet
conncetivity.