Big cleanup in MessageHandler Registry code.
Also restructured other parts of the code. A lot of fake singletons got real
singletons. Some renamings may not be ideal but they are easy to
refactor later.

All in all, following the needs of #24 in order to make main.cpp smaller
so it is easier for customers and users to implement their own ideas in
a short and understandable main.cpp.