webserver: Static File Staging and Serving.
First implementation, however new problems appear while implementing it.
For instance, the aWOT lib makes probably more problems then it solves
(it is almost no more maintained). The bug which is currently noted in
the code is that the return value of Stream.write(byte*, size_t size) is not
properly passed back and therefore it is impossible to fully write
slightly larger replies. Easy to fix, but requires toughing aWOT. Bug will
be tracked in a seperate issue.