Adding an easer egg #123
Code compiles but not yet tested. Needs most likely an adaption for REV1 hardware. The Carrier access #124 is supposed to be provided throught a message handler. The location of the code in the mode directory is because this is supposed to be a hidden mode in some global state machine #112.