Improved logging.
We now can log as JSONL to different channels (such as Serial or an Ethernet client) at the same time. Furthermore, we can log to RAM, in particular as a startup log #52. In order to do this somewhat efficiently, we introduced JSON streaming #122 (kind of JSON SAX  instead of JSON DOM as ArduinoJSON does it).

Code compiles but not yet tested (same with previous commit).