Fixing startup bug.
The init() function never completed because the bus::init() was missing. The REV1->lucidac merge is now finally workable. This commit still contains debugging code in the main file (hybrid_controller.cpp).

For the issue whether Carrier{0} vs LUCIDAC shall be used see also #138 but in particular the singleton question #124.