Get_entities hands out EUI-64.
This solves #143. It would be nicer if the code would read out all of the EEPROMs at startup and the FunctionBlocks hold that value. However, the FrontPanel is a FunctionBlock (it should be one, IMHO). And the need was simpler fulfiled this way anyway. Code compiles and tested on real hardware.