Docs: First step towards better docs #210
Sphinx is just much better readable compared to Doxygen, when it comes
to the non-API part. The API is much better covered by Doxygen, thought.
In the future, we just will have both, despite the Doxygen part will be
more focussed on the API and the Sphinx on the rest.